create strava etl class
