+++ fitnessllm_dataplatform/services/etl_interface.py
@@ -1,0 +1,3 @@
+class ETLInterface:
+    def __init__(self):
+        pass
